+++ internal/handlers/routes.go
@@ -280,8 +280,8 @@ func InitRoutes(app *core.App, fiberApp *fiber.App) {
 	v1Manga.Get("/entry/:id/details", makeHandler(app, HandleGetMangaEntryDetails))
 	v1Manga.Post("/chapters", makeHandler(app, HandleGetMangaEntryChapters))
 	v1Manga.Post("/pages", makeHandler(app, HandleGetMangaEntryPages))
-	v1Manga.Post("/entry/backups", makeHandler(app, HandleGetMangaEntryBackups))
-	v1Manga.Post("/download-chapter", makeHandler(app, HandleDownloadMangaChapter))
+	//v1Manga.Post("/entry/backups", makeHandler(app, HandleGetMangaEntryBackups))
+	//v1Manga.Post("/download-chapter", makeHandler(app, HandleDownloadMangaChapter))
 	v1Manga.Post("/update-progress", makeHandler(app, HandleUpdateMangaProgress))
 
 	//