+++ seanime-web/src/components/application/beta-badge.tsx
@@ -4,6 +4,6 @@ type Props = BadgeProps
 
 export function BetaBadge(props: Props) {
     return (
-        <Badge intent="warning" className="align-middle ml-1.5" {...props}>Experimental</Badge>
+        <Badge intent="warning" size="sm" className="align-middle ml-1.5" {...props}>Beta</Badge>
     )
 }